+++ common/src/main/java/com/traverse/taverntokens/item/BagItemTooltip.java
@@ -1,0 +1,18 @@
+package com.traverse.taverntokens.item;
+
+import com.traverse.taverntokens.wallet.WalletItemStack;
+
+import net.minecraft.core.NonNullList;
+import net.minecraft.world.inventory.tooltip.TooltipComponent;
+
+public class BagItemTooltip implements TooltipComponent {
+    private final NonNullList<WalletItemStack> items;
+
+    public BagItemTooltip(NonNullList<WalletItemStack> nonNullList) {
+        this.items = nonNullList;
+    }
+
+    public NonNullList<WalletItemStack> getItems() {
+        return this.items;
+    }
+}