+++ common/src/main/java/com/traverse/taverntokens/mixin/GuiGraphicsMixin.java
@@ -1,5 +1,9 @@
 package com.traverse.taverntokens.mixin;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -9,10 +13,17 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import com.traverse.taverntokens.item.BagItemTooltip;
+import com.traverse.taverntokens.item.ClientBagItemTooltip;
 import com.traverse.taverntokens.wallet.WalletItemStack;
 
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipPositioner;
+import net.minecraft.client.gui.screens.inventory.tooltip.DefaultTooltipPositioner;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.ItemStack;
 
 @Mixin(GuiGraphics.class)
@@ -25,14 +36,18 @@ public abstract class GuiGraphicsMixin {
     @Shadow
     public abstract int drawString(Font font, String text, int x, int y, int color, boolean shadow);
 
+    @Shadow
+    protected abstract void renderTooltipInternal(Font font, List<ClientTooltipComponent> list2, int x, int y,
+            ClientTooltipPositioner instance);
+
     @ModifyVariable(at = @At("HEAD"), method = "renderItemDecorations(Lnet/minecraft/client/gui/Font;Lnet/minecraft/world/item/ItemStack;IILjava/lang/String;)V")
-    private String renderItemDecorations$ModifyVariable(String countOverride, Font font, ItemStack stack,
+    private String renderItemDecorations(String countOverride, Font font, ItemStack stack,
             int x, int y) {
         return (stack instanceof WalletItemStack) ? "" : countOverride;
     }
 
     @Inject(at = @At("HEAD"), method = "renderItemDecorations(Lnet/minecraft/client/gui/Font;Lnet/minecraft/world/item/ItemStack;IILjava/lang/String;)V")
-    private void renderItemDecorations$Inject(Font font, ItemStack stack, int x, int y, String countOverride,
+    private void renderItemDecorations(Font font, ItemStack stack, int x, int y, String countOverride,
             CallbackInfo info) {
         if (!stack.isEmpty()) {
             if (stack instanceof WalletItemStack walletItemStack) {
@@ -51,5 +66,22 @@ public abstract class GuiGraphicsMixin {
                 }
             }
         }
+    }
+
+    @Inject(method = "Lnet/minecraft/client/gui/GuiGraphics;renderTooltip(Lnet/minecraft/client/gui/Font;Ljava/util/List;Ljava/util/Optional;II)V", at = @At("HEAD"), cancellable = true)
+    private void renderTooltip(Font font, List<Component> list, Optional<TooltipComponent> optional, int x, int y,
+            CallbackInfo ci) {
+
+        optional.ifPresent((component) -> {
+            if (optional.get() instanceof BagItemTooltip tooltip) {
+                List<ClientTooltipComponent> list2 = list.stream().map(Component::getVisualOrderText)
+                        .map(ClientTooltipComponent::create).collect(Collectors.toList());
+
+                list2.add(1, new ClientBagItemTooltip(tooltip));
+
+                this.renderTooltipInternal(font, list2, x, y, DefaultTooltipPositioner.INSTANCE);
+                ci.cancel();
+            }
+        });
     }
 }