+++ forge/src/main/java/com/traverse/taverntokens/registry/ForgeModItems.java
@@ -1,13 +1,13 @@
 package com.traverse.taverntokens.registry;
 
 import com.traverse.taverntokens.ForgeTavernTokens;
+import com.traverse.taverntokens.item.BagItem;
 
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
@@ -21,7 +21,7 @@ public class ForgeModItems {
             ForgeTavernTokens.MODID);
 
     public static final RegistryObject<Item> MONEY_BAG_REGISTRY = ITEMS.register("money_bag",
-            () -> new Item(new Item.Properties()));
+            () -> new BagItem(new Item.Properties()));
     public static final RegistryObject<Item> COPPER_COIN_REGISTRY = ITEMS.register("copper_coin",
             () -> new Item(new Item.Properties()));
     public static final RegistryObject<Item> IRON_COIN_REGISTRY = ITEMS.register("iron_coin",
@@ -34,7 +34,7 @@ public class ForgeModItems {
     public static RegistryObject<CreativeModeTab> CREATIVE_TAB = TAB.register("taverntokens_item_group",
             () -> CreativeModeTab.builder()
                     .title(Component.translatable("itemGroup." + ForgeTavernTokens.MODID))
-                    .icon(() -> new ItemStack(Items.BUNDLE))
+                    .icon(() -> new ItemStack(MONEY_BAG_REGISTRY.get()))
                     .displayItems((itemDisplayParameters, output) -> {
                         output.accept(MONEY_BAG_REGISTRY.get());
                         output.accept(COPPER_COIN_REGISTRY.get());